Change elapse to float

+++ leo-gps.c
@@ -1010,13 +1010,13 @@ static void* gps_timer_thread( void*  arg ) {
 
         if (r->fix_flags_cached) {
             clock_t now = clock();
-            int elapsed = 1;
+            float elapsed = 1;
             int fix_freqflag=1;
             D("fix_freqflag = 1");
             do{
                 usleep((uint64_t)500000);
                 elapsed = (clock()-now)/CLOCKS_PER_SEC;
-                D("elapsed = %d",elapsed);
+                D("elapsed = %f",elapsed);
 		if(fix_temp > 3 && elapsed<5)
                 {
                     state->fix_freq = 1;
